fix(sonar): 불필요한 double 캐스트 제거 (java:S1905)

+++ src/main/java/com/backend/onharu/utils/NumberUtils.java
@@ -19,7 +19,7 @@ public final class NumberUtils {
         if (value == null || Double.isNaN(value)) {
             return 0.0;
         }
-        return (double) value.longValue();
+        return value.longValue();
     }
 
     /**
